refactor: split Navbar into NavLinks and NavWalletStatus components (#183)
- Extract NavLinks and NavMobileLinks into src/components/navbar/NavLinks.tsx
- Extract NavWalletStatus into src/components/navbar/NavWalletStatus.tsx
- Navbar.tsx becomes a slim shell wiring the sub-components together

Closes #183
